Extract the story editor to a module
