+++ lib/modules/story_editor/widgets/narrative_list_view.dart
@@ -3,11 +3,11 @@ import 'package:andax/models/node.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/models/translation.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/widgets/modal_picker.dart';
+import 'package:andax/modules/story_editor/widgets/modal_picker.dart';
 import 'package:andax/widgets/rounded_back_button.dart';
 import 'package:flutter/material.dart';
 
-import 'story_editor_screen.dart';
+import '../screens/narrative_editor.dart';
 
 class NarrativeListView extends StatefulWidget {
   const NarrativeListView(
